switched indexWeight to long

+++ kmp/src/commonMain/kotlin/pl/techbrewery/sam/kmp/repository/ShoppingListRepository.kt
@@ -7,7 +7,7 @@ import pl.techbrewery.sam.kmp.database.entity.SingleItem
 class ShoppingListRepository(
     private val db: KmpDatabase
 ) {
-    suspend fun insertItem(itemName: String, indexWeight: Float) {
+    suspend fun insertItem(itemName: String, indexWeight: Long) {
         val item = SingleItem(
             itemName = itemName.lowercase(),
             indexWeight = indexWeight
